fix: cleaning up files for programmatic view
