+++ Sources/GDSCommon/UI/UIKit/SharedViews/Patterns/LinkTheApp/LinkTheApp.xib
@@ -16,9 +16,6 @@
                 <outlet property="firstSectionSubtitleLabel" destination="0tU-u6-faS" id="vPi-ow-hOU"/>
                 <outlet property="firstSectionTitleLabel" destination="rYa-6N-pGl" id="3o9-a2-dHL"/>
                 <outlet property="imageView" destination="JEv-PL-uNu" id="bQ2-Wc-oCR"/>
-                <outlet property="secondSectionButton" destination="96X-lT-Dwj" id="KTs-jJ-88e"/>
-                <outlet property="secondSectionSubtitleLabel" destination="MJ7-SW-trK" id="Ftz-cC-0Qy"/>
-                <outlet property="secondSectionTitleLabel" destination="acD-Le-7rw" id="ESX-NX-VBS"/>
                 <outlet property="subtitleLabel" destination="urP-GV-7kQ" id="eJb-Jm-SPg"/>
                 <outlet property="titleLabel" destination="FdJ-8g-DNa" id="Bfb-My-cne"/>
                 <outlet property="view" destination="iN0-l3-epB" id="wo3-gP-a0g"/>
@@ -33,7 +30,7 @@
                     <rect key="frame" x="0.0" y="59" width="393" height="793"/>
                     <subviews>
                         <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="16" translatesAutoresizingMaskIntoConstraints="NO" id="ja5-PX-D17" userLabel="Content View">
-                            <rect key="frame" x="0.0" y="24.999999999999972" width="393" height="477.33333333333326"/>
+                            <rect key="frame" x="0.0" y="25" width="393" height="346.66666666666669"/>
                             <subviews>
                                 <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="16" translatesAutoresizingMaskIntoConstraints="NO" id="dp2-Mq-JlH">
                                     <rect key="frame" x="0.0" y="0.0" width="393" height="216"/>
@@ -89,42 +86,6 @@
                                             <constraints>
                                                 <constraint firstAttribute="height" relation="greaterThanOrEqual" constant="20" id="TMq-i1-OsZ"/>
                                                 <constraint firstAttribute="height" relation="lessThanOrEqual" constant="44" id="aSW-kh-jeg"/>
-                                            </constraints>
-                                            <fontDescription key="fontDescription" style="UICTFontTextStyleBody"/>
-                                            <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
-                                            <state key="normal" title="Button"/>
-                                        </button>
-                                    </subviews>
-                                    <edgeInsets key="layoutMargins" top="8" left="16" bottom="8" right="16"/>
-                                </stackView>
-                                <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="8" translatesAutoresizingMaskIntoConstraints="NO" id="xBH-Cp-NJ6">
-                                    <rect key="frame" x="0.0" y="362.66666666666669" width="393" height="114.66666666666669"/>
-                                    <subviews>
-                                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" text="Label" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontForContentSizeCategory="YES" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="acD-Le-7rw">
-                                            <rect key="frame" x="16" y="7.9999999999999982" width="361" height="20.333333333333329"/>
-                                            <fontDescription key="fontDescription" style="UICTFontTextStyleHeadline"/>
-                                            <nil key="textColor"/>
-                                            <nil key="highlightedColor"/>
-                                        </label>
-                                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" text="Label" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontForContentSizeCategory="YES" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="MJ7-SW-trK">
-                                            <rect key="frame" x="16" y="36.333333333333314" width="361" height="20.333333333333329"/>
-                                            <fontDescription key="fontDescription" style="UICTFontTextStyleBody"/>
-                                            <nil key="textColor"/>
-                                            <nil key="highlightedColor"/>
-                                        </label>
-                                        <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="WbC-8e-nbv">
-                                            <rect key="frame" x="16" y="64.666666666666629" width="361" height="1"/>
-                                            <color key="backgroundColor" white="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                            <color key="tintColor" white="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                            <constraints>
-                                                <constraint firstAttribute="height" constant="1" id="tfx-pb-BYN"/>
-                                            </constraints>
-                                        </view>
-                                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="leading" contentVerticalAlignment="center" buttonType="system" lineBreakMode="wordWrap" translatesAutoresizingMaskIntoConstraints="NO" id="96X-lT-Dwj">
-                                            <rect key="frame" x="16" y="73.666666666666686" width="361" height="33"/>
-                                            <constraints>
-                                                <constraint firstAttribute="height" relation="lessThanOrEqual" constant="44" id="1uL-4L-IAs"/>
-                                                <constraint firstAttribute="height" relation="greaterThanOrEqual" constant="20" id="gdY-sa-6f8"/>
                                             </constraints>
                                             <fontDescription key="fontDescription" style="UICTFontTextStyleBody"/>
                                             <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
