+++ Sources/GDSCommon/UI/UIKit/SharedViews/Patterns/LinkTheApp/LinkTheAppViewModel.swift
@@ -3,6 +3,6 @@ import UIKit
 public protocol LinkTheAppViewModel {
     var imageName: String { get }
     var title: GDSLocalisedString { get }
-    var subtitle: GDSLocalisedString { get }
+    var body: GDSLocalisedString { get }
     var contentViews: [UIView] { get }
 }
